renamed calculate fee oracle to dynamic fee
