+++ packages/sdk/src/chains/EVM/fee/index.ts
@@ -1,3 +1,3 @@
-export { calculateFeeData } from './feeOracle';
+export { calculateDynamicFee as calculateFeeData } from './dynamicFee';
 export { calculateBasicfee } from './basicFee';
 export { getFeeHandlerAddress } from './feeHandler';
